codellama-project/backend/requirements.txt
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.44.2
fastapi==0.103.0
uvicorn==0.23.2
accelerate==0.33.0
bitsandbytes==0.41.0
psycopg2-binary==2.9.9
pydantic==2.8.2